Avoid forward references in `commitUtils` types
Summary:
The resulting types less logically structured (as discussed), but while
Flow considers them equivalent, the linter does not—and, more
importantly, the runtime `PropTypes` are not generated correctly when
forward-uses are in play, due to the following issue:

https://github.com/brigand/babel-plugin-flow-react-proptypes/issues/185

I _think_ this is the lesser of two evils—but am happy to be convinced
otherwise.

Test Plan:
Note that `yarn flow` and `yarn test` are both fine, and `yarn start`
has no lint errors or runtime errors. Note further that when reverting
`insertions → added` and `deletions → deleted`, there are prop-type
errors in both `yarn test` and `yarn start`. (Without the permutation of
the types, this does not occur.)

wchargin-branch: avoid-forward-references