Remove outdated design documents (#45)
design.md and overview.md both describe a vision of sourcecred
in which it is a measure of credit, and an explicit cryptographic
token. right now, SourceCred is more focused on just measuring credit,
with the expectation that cryptoincentives can be added on later.

Removing these outdated documents will reduce confusion; they may
be re-written and re-added later.