Convert ArtifactList to ArtifactGraphEditor
Summary:
This component now maintains a graph of just artifacts and the edges
among them. It owns the state, and notifies its parents of changes with
a callback. We treat the graph objects as properly immutable, copying
them on each change. So far, descriptions are always the empty string.

Test Plan:
Interaction tests added. Run `yarn test`.

wchargin-branch: artifact-graph-editor