GitHub: parse simple numeric references
This commit adds support for parsing numeric references to the GitHub
plugin. The implementation is mildly complicated, because the parser
now tracks two additional pieces of state:

1. A map from issue/pr # to node, necessary to avoid an O(nodes) scan
every time we encounter a numeric reference
2. A map containing every dangling numeric reference, necessary to avoid
missing references when we parse the reference before the referred node.

I also added a `danglingReferences` method to the parser which
enumerates every dangling reference, in a readable format. In a followon
commit, I intend to log the presence of dangling references in the
`parseGraph` method of the Artifact editor.

Test plan:
Observe that the unit tests are fairly comprehensive. The whole-repo
snapshot changes are as expected.