Artifact plugin logs presence of dangling edges
If there are any dangling edges leftover from a graph parse, the
artifact plugin will now record that fact in the console.

I think this will be mildly useful for debugging; note it only generates
noise in the exception case.

Test plan:
No tests were added. We don't depend on this behavior at all, I just
think it will be nice to make it discoverable when this happens.